requirements.txt
Django>=4.2
djangorestframework
celery
redis
transformers
torch
PyYAML